+++ src/Output/Output.ts
@@ -1,3 +1,5 @@
+import { Stringable } from '../Stringable/Stringable.ts'
+
 /* eslint @typescript-eslint/no-explicit-any: 0 -- Params can be anything. */
 
 /**
@@ -11,48 +13,48 @@ export interface Output {
 	 *
 	 * @since 0.1.0
 	 *
-	 * @param {string} message           Message to output.
-	 * @param {any[]}  ...optionalParams Optional parameters.
+	 * @param {string | Stringable} message           Message to output.
+	 * @param {any[]}               ...optionalParams Optional parameters.
 	 */
-	error(message: string, ...optionalParams: any[]): void
+	error(message: string | Stringable, ...optionalParams: any[]): void
 
 	/**
 	 * Prints a warning to an error stream.
 	 *
 	 * @since 0.1.0
 	 *
-	 * @param {string} message           Message to output.
-	 * @param {any[]}  ...optionalParams Optional parameters.
+	 * @param {string | Stringable} message           Message to output.
+	 * @param {any[]}               ...optionalParams Optional parameters.
 	 */
-	warn(message: string, ...optionalParams: any[]): void
+	warn(message: string | Stringable, ...optionalParams: any[]): void
 
 	/**
 	 * Prints a log message to an output stream.
 	 *
 	 * @since 0.1.0
 	 *
-	 * @param {string} message           Message to output.
-	 * @param {any[]}  ...optionalParams Optional parameters.
+	 * @param {string | Stringable} message           Message to output.
+	 * @param {any[]}               ...optionalParams Optional parameters.
 	 */
-	log(message: string, ...optionalParams: any[]): void
+	log(message: string | Stringable, ...optionalParams: any[]): void
 
 	/**
 	 * Prints an info message to an output stream.
 	 *
 	 * @since 0.1.0
 	 *
-	 * @param {string} message           Message to output.
-	 * @param {any[]}  ...optionalParams Optional parameters.
+	 * @param {string | Stringable} message           Message to output.
+	 * @param {any[]}               ...optionalParams Optional parameters.
 	 */
-	info(message: string, ...optionalParams: any[]): void
+	info(message: string | Stringable, ...optionalParams: any[]): void
 
 	/**
 	 * Print a debug message to an output stream.
 	 *
 	 * @since 0.1.0
 	 *
-	 * @param {string} message           Message to output.
-	 * @param {any[]}  ...optionalParams Optional parameters.
+	 * @param {string | Stringable} message           Message to output.
+	 * @param {any[]}               ...optionalParams Optional parameters.
 	 */
-	debug(message: string, ...optionalParams: any[]): void
+	debug(message: string | Stringable, ...optionalParams: any[]): void
 }