+++ src/Stringable/Stringable.ts
@@ -1,0 +1,15 @@
+/**
+ * Object which implements toString method.
+ *
+ * @since unreleased
+ */
+export interface Stringable {
+	/**
+	 * Converts object to a string.
+	 *
+	 * @since  unreleased
+	 *
+	 * @return {string}
+	 */
+	toString(): string
+}