+++ CHANGELOG.md
@@ -16,7 +16,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   Stringable interface.
+-   MessageStringable class.
+
 ### Changed
+
+-   Allow Output methods' first parameter to be Stringable.
 
 ### Deprecated